requirements.txt
Django==4.1.4
django-extensions~=3.2.1
python-dateutil~=2.8.2
django-countries==7.5
django-versatileimagefield~=2.2
djangorestframework~=3.14.0
django-filter~=22.1
djangorestframework-simplejwt~=5.2.2